backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
google-generativeai==0.7.2
gTTS==2.5.1
python-multipart==0.0.9
httpx==0.27.0
